levelup_project/order/serializers.py
from django.db.models import fields
from rest_framework import serializers
from order.models import Shop, Menu, Order, Order_food

class ShopSerializer(serializers.ModelSerializer):
    class Meta:
        model = Shop
        fields = '__all__'

class MenuSerializer(serializers.ModelSerializer):
    class Meta:
        model = Menu
        fields = '__all__'