levelup_project/requirements.txt
asgiref==3.4.1
Django==4.0
djangorestframework==3.13.1
mysqlclient==2.1.0
protobuf==3.19.1
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
